pyproject.toml
[tool.poetry]
name = "postgres-homeworks"
version = "0.1.0"
description = ""
authors = ["LLiRRiK <kill2002@mail.ru>"]
readme = "README.md"
packages = [{include = "postgres_homeworks"}]

[tool.poetry.dependencies]
python = "^3.11"
psycopg2 = "^2.9.7"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
